+++ src/shared/BackgroundModel.h
@@ -20,18 +20,18 @@ class BackgroundModel{
 public:
 
 	BackgroundModel( SequenceSet& sequenceSet,
-			         int order,
-			         std::vector<float> alpha,
-			         bool interpolate = true,
-			         std::vector< std::vector<int> > foldIndices = std::vector< std::vector<int> >(),
-			         std::vector<int> folds = std::vector<int>() );
+					size_t order,
+			        std::vector<float> alpha,
+			        bool interpolate = true,
+			        std::vector<std::vector<size_t>> foldIndices = std::vector<std::vector<size_t>>(),
+			        std::vector<size_t> folds = std::vector<size_t>() );
 
 	BackgroundModel( std::string filePath );
 
 	~BackgroundModel();
 
 	std::string getName();
-	int			getOrder();
+	size_t		getOrder();
 	float**		getV();
 
 	void 		expV();
@@ -41,16 +41,16 @@ public:
 
 	// calculate the log likelihood for the sequence set
 	// afterwards v_ contains log probabilities
-	double calculateLogLikelihood( SequenceSet& sequenceSet,
-			                       std::vector<std::vector<int>> foldIndices = std::vector<std::vector<int>>(),
-			                       std::vector<int> folds = std::vector<int>() );
+	double 	calculateLogLikelihood( SequenceSet& sequenceSet,
+			                       std::vector<std::vector<size_t>> foldIndices = std::vector<std::vector<size_t>>(),
+			                       std::vector<size_t> folds = std::vector<size_t>() );
 
 	// calculate positional likelihoods for the sequence set
 	// and write likelihoods to file
-	void calculatePosLikelihoods( SequenceSet& sequenceSet,
-			                      char* outputDirectory,
-			                      std::vector<std::vector<int>> foldIndices = std::vector<std::vector<int>>(),
-			                      std::vector<int> folds = std::vector<int>() );
+	void 	calculatePosLikelihoods( SequenceSet& sequenceSet,
+			                      	  char* outputDirectory,
+			                      	  std::vector<std::vector<size_t>> foldIndices = std::vector<std::vector<size_t>>(),
+			                      	  std::vector<size_t> folds = std::vector<size_t>() );
 
 
 	void 	print();
@@ -66,18 +66,18 @@ private:
 
 	std::string			name_;					// basename of sequence set file
 
-	int**				n_;						// oligomer counts
+	size_t**			n_;						// oligomer counts
 	float** 			v_;						// oligomer conditional probabilities
 
 	bool				vIsLog_ = false;		// v_ contains log probabilities
 
-	int					K_;						// order
+	size_t				K_;						// order
 	std::vector<float>	A_;						// order-specific alphas
 
 	bool				interpolate_ = true;	// calculate prior probabilities from lower-order probabilities
 												// instead of background frequencies of mononucleotides
 
-	std::vector<int>	Y_;						// contains 1 at position 0
+	std::vector<size_t>	Y_;						// contains 1 at position 0
 												// and the number of oligomers y for increasing order k (from 0 to K_) at positions k+1
 												// e.g. alphabet size_ = 4 and K_ = 2: Y_ = 1 4 16 64
 };
